+++ src/main/java/com/example/backendmainserver/client/raspberry/dto/request/PortAndSupplier.java
@@ -1,4 +1,4 @@
-package com.example.backendmainserver.client.raspberry.dto;
+package com.example.backendmainserver.client.raspberry.dto.request;
 
 
 public record PortAndSupplier(Long portId, String powerSupplier) {
